fix(desktop): preserve authority-form local file links
Problem: plain-text authority-form UNC file URLs were inert, and exported Markdown dropped their href because allowlists only recognized file:/// URLs.

Root cause: local-path linkification used a file:/// only matcher, while Markdown rendering, click handling, and export each carried separate file URL validation rules.

Fix: centralize strict file URL parsing with the standard URL API, recognize canonical and slash-form UNC URLs, reject malformed candidates before linkification, preserve valid file URLs when generating anchors, and reuse the validator for Markdown and export sanitizers.

Verification: local-path tests 61/61; click/context-menu E2E 17/17; session export tests; pnpm test (120 suites); pnpm typecheck; frontend hooks lint and CSS checks; frontend production build and bundle budgets; cd desktop && go test ./...; go test -race ./...; go vet ./... .
